backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.31
gspread==6.1.2
google-auth==2.30.0
python-frontmatter==1.1.0
markdown==3.6
pyyaml==6.0.1
python-multipart==0.0.9
aiofiles==23.2.1
